+++ src/main/java/de/markusbordihn/playercompanions/commands/CommandManager.java
@@ -51,6 +51,7 @@ public class CommandManager {
         .then(BackupCommand.register())
         .then(ListCommand.register())
         .then(OverviewCommand.register())
+        .then(SummonCommand.register())
         .then(VersionCommand.register())
       // @formatter:on
     );