+++ src/main/java/de/markusbordihn/playercompanions/network/message/MessageCommandPlayerCompanion.java
@@ -65,6 +65,10 @@ public class MessageCommandPlayerCompanion {
   public static void handlePacket(MessageCommandPlayerCompanion message,
       NetworkEvent.Context context) {
     ServerPlayer serverPlayer = context.getSender();
+    if (serverPlayer == null) {
+      log.error("Unable to get server player for message {} from {}", message, context);
+      return;
+    }
     ServerLevel serverLevel = serverPlayer.getLevel();
     UUID uuid = UUID.fromString(message.getPlayerCompanionUUID());
     Entity entity = serverLevel.getEntity(uuid);